adds /library page
