+++ lyric-keeper-web-app/src/Homepage/Homepage.tsx
@@ -1,60 +1,35 @@
 import React, { useEffect, useState } from "react";
 import {
   WelcomeText,
-  DefaultPageWrapper,
-  MainAreaWrapper,
-  StyledSelect,
   StyledSwitch,
-  StyledTextField,
-  StyledIconButton,
   DarkIcon,
   LightIcon,
-  NoLyricsToDisplayText,
   StyledLink,
 } from "./elements";
-import { LyricCard } from "LyricCard";
 import { NewLyricModal } from "NewLyricModal";
-import { useQuery, useMutation } from "react-apollo";
-import MenuItem from "@material-ui/core/MenuItem";
-import RefreshIcon from "@material-ui/icons/Refresh";
+import { useMutation } from "react-apollo";
 import HelpOutlineIcon from "@material-ui/icons/HelpOutline";
 import {
-  Query_Get_All_Lyrics,
   Mutation_Add_New_Lyric,
   Mutation_Add_New_Lyric_To_User_List,
 } from "operations";
 import {
-  Get_All_Lyrics,
   Lyric,
   Add_New_LyricVariables,
   Add_New_Lyric_To_User_List_addNewLyricToUserList,
   Add_New_Lyric_To_User_ListVariables,
   Add_New_Lyric,
-  Get_All_Lyrics_allLyrics,
 } from "Types";
-import { useFormik } from "formik";
 import {
   LoginCreateAccountModal,
   MarketingModal,
   MarketingBar,
-  LoadingScreen,
   Navbar,
-  LyricCountWrapper,
+  DefaultPageWrapper,
 } from "GlobalComponents";
 import { UseCurrentUser, UseDarkMode } from "Hooks";
-import { findNonPrivateLyrics } from "utilities";
 
-type allLyrics = Get_All_Lyrics["allLyrics"];
-
-export interface SettingsObj {
-  refetchLyrics?: boolean;
-}
-
-export const Homepage: React.FC<any> = ({ client }) => {
-  const [lyricDataSourceOfTruth, setLyricDataSourceOfTruth] = useState<
-    allLyrics | undefined
-  >([]);
-  const [lyricData, setLyricData] = useState<allLyrics | undefined>([]);
+export const Homepage: React.FC = () => {
   const [loginModalIsOpen, setLoginModalIsOpen] = useState(false);
   const [marketingModalIsOpen, setMarketingModalIsOpen] = useState(true);
 
@@ -69,16 +44,6 @@ export const Homepage: React.FC<any> = ({ client }) => {
 
   const { darkModeIsEnabled, setDarkMode } = UseDarkMode();
 
-  const {
-    handleChange,
-    values: { search, filterBy },
-  } = useFormik({
-    initialValues: { search: "", filterBy: "title" as "title" | "author" },
-    onSubmit: () => undefined,
-  });
-
-  const { data, loading, refetch } = useQuery(Query_Get_All_Lyrics);
-
   const [addNewLyric] = useMutation<Add_New_Lyric, Add_New_LyricVariables>(
     Mutation_Add_New_Lyric
   );
@@ -87,45 +52,6 @@ export const Homepage: React.FC<any> = ({ client }) => {
     Add_New_Lyric_To_User_List_addNewLyricToUserList,
     Add_New_Lyric_To_User_ListVariables
   >(Mutation_Add_New_Lyric_To_User_List);
-
-  const filter = (searchTerm: string) => {
-    setLyricData(
-      lyricDataSourceOfTruth &&
-        lyricDataSourceOfTruth.filter(item =>
-          item[filterBy].toLowerCase().includes(searchTerm.toLowerCase())
-        )
-    );
-  };
-
-  const getAndUpdateAllLyrics = (settings?: SettingsObj) => {
-    settings?.refetchLyrics && refetch();
-    // Handled offline data
-    if (!data || loading) {
-      try {
-        const cachedData = client.readQuery({
-          query: Query_Get_All_Lyrics,
-        });
-        const nonPrivateLyrics = findNonPrivateLyrics<Get_All_Lyrics_allLyrics>(
-          cachedData.allLyrics
-        );
-        setLyricData(nonPrivateLyrics);
-        setLyricDataSourceOfTruth(nonPrivateLyrics);
-      } catch (error) {
-        console.warn(error);
-      }
-    } else if (data && !loading) {
-      const nonPrivateLyrics = findNonPrivateLyrics<Get_All_Lyrics_allLyrics>(
-        data?.allLyrics
-      );
-      setLyricDataSourceOfTruth(nonPrivateLyrics);
-      setLyricData(nonPrivateLyrics);
-    }
-  };
-
-  useEffect(() => {
-    getAndUpdateAllLyrics({ refetchLyrics: true });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data, loading]);
 
   useEffect(() => {
     if (
@@ -142,7 +68,7 @@ export const Homepage: React.FC<any> = ({ client }) => {
   const addEntry = async (lyric: Lyric) => {
     if (currentUser) {
       const lyricData = await addNewLyric({ variables: lyric });
-      getAndUpdateAllLyrics({ refetchLyrics: true });
+      // getAndUpdateAllLyrics({ refetchLyrics: true });
 
       const lyricId = lyricData.data?.addNewLyric?.id;
 
@@ -159,13 +85,12 @@ export const Homepage: React.FC<any> = ({ client }) => {
     setLoginModalIsOpen(true);
   };
 
-  // if (loading) return <LoadingScreen darkMode={darkModeIsEnabled} />;
-
   return (
     <>
       {!isLoggedIn && (
         <MarketingBar onLoginButtonClick={() => setLoginModalIsOpen(true)} />
       )}
+      {isLoggedIn && <NewLyricModal addEntry={addEntry} />}
       <Navbar {...currentUserDetails} />
       <DefaultPageWrapper darkMode={darkModeIsEnabled}>
         <LoginCreateAccountModal
@@ -193,7 +118,7 @@ export const Homepage: React.FC<any> = ({ client }) => {
           checked={darkModeIsEnabled}
           onChange={({ target: { checked } }) => setDarkMode(checked)}
         />
-        <MainAreaWrapper maxWidth="sm">
+        {/* <MainAreaWrapper maxWidth="sm">
           <StyledTextField
             darkMode={darkModeIsEnabled}
             label="Search"
@@ -245,7 +170,7 @@ export const Homepage: React.FC<any> = ({ client }) => {
             </>
           )}
           {isLoggedIn && <NewLyricModal addEntry={addEntry} />}
-        </MainAreaWrapper>
+        </MainAreaWrapper> */}
       </DefaultPageWrapper>
     </>
   );
