+++ lyric-keeper-web-app/src/Homepage/LoggedInView.tsx
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const LoggedInView: React.FC = () => <></>;
