+++ lyric-keeper-web-app/src/Homepage/LoggedOutView.tsx
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const LoggedOutView: React.FC = () => <></>;
