+++ lyric-keeper-web-app/src/Homepage/elements.tsx
@@ -26,31 +26,6 @@ export const WelcomeText = styled(Typography)<{ darkMode: boolean }>`
   }
 `;
 
-export const DefaultPageWrapper = styled.div<{ darkMode: boolean }>`
-  text-align: center;
-  background-color: ${({ darkMode }) => (darkMode ? `#1a1a1a` : `white`)};
-  padding-top: 30px;
-  min-height: 88vh;
-`;
-
-export const MainAreaWrapper = styled(Container)`
-  margin: auto;
-  padding-bottom: 70px;
-`;
-
-export const StyledSelect = styled(Select)<{ darkMode: boolean }>`
-  width: 4.5rem;
-  display: inline-block;
-  margin-top: 16px;
-  left: 13px;
-  color: ${({ darkMode }) => (darkMode ? `white` : `black`)};
-  & {
-    .MuiSelect-icon {
-      color: ${({ darkMode }) => (darkMode ? `white` : `black`)};
-    }
-  }
-`;
-
 export const StyledSwitch = styled(Switch)`
   && {
     .MuiSwitch-colorSecondary.Mui-checked {
@@ -60,21 +35,6 @@ export const StyledSwitch = styled(Switch)`
       background-color: grey;
     }
   }
-`;
-
-export const StyledTextField = styled(TextField)<{ darkMode: boolean }>`
-  & {
-    .MuiInputBase-input {
-      color: ${({ darkMode }) => (darkMode ? `white` : `black`)};
-    }
-    .MuiInputLabel-root {
-      color: ${({ darkMode }) => (darkMode ? `white` : `black`)};
-    }
-  }
-`;
-
-export const StyledIconButton = styled(IconButton)<{ darkMode: boolean }>`
-  color: ${({ darkMode }) => (darkMode ? `white` : `black`)};
 `;
 
 export const LightIcon = styled(WbSunnyIcon)`
@@ -95,14 +55,6 @@ export const DarkIcon = styled(NightsStayIcon)`
     margin-bottom: -20px;
     font-size: 40px;
   }
-`;
-
-export const NoLyricsToDisplayText = styled.div<{ darkMode: boolean }>`
-  margin-top: 5vh;
-  font-size: 3vh;
-  display: block;
-  letter-spacing: 0.2vw;
-  color: ${({ darkMode }) => (darkMode ? "white" : "black")};
 `;
 
 export const StyledLink = styled(Link)`
